+++ src/OptionsBuilder/BirthdayOptionsBuilder.php
@@ -20,6 +20,7 @@ class BirthdayOptionsBuilder extends AbstractOptionsBuilder
     public const WIDGET_TEXT = 'text';
     public const WIDGET_SINGLE_TEXT = 'single_text';
 
+    /** @return class-string */
     public static function getBuilderType(): string
     {
         return BirthdayType::class;