+++ src/OptionsBuilder/CheckboxOptionsBuilder.php
@@ -12,6 +12,7 @@ use Symfony\Component\Form\Extension\Core\Type\CheckboxType;
 
 class CheckboxOptionsBuilder extends AbstractOptionsBuilder
 {
+    /** @return class-string */
     public static function getBuilderType(): string
     {
         return CheckboxType::class;