+++ src/OptionsBuilder/HiddenOptionsBuilder.php
@@ -9,6 +9,7 @@ use Symfony\Component\Form\Extension\Core\Type\HiddenType;
 
 class HiddenOptionsBuilder extends AbstractOptionsBuilder
 {
+    /** @return class-string */
     public static function getBuilderType(): string
     {
         return HiddenType::class;