+++ src/OptionsBuilder/PasswordOptionsBuilder.php
@@ -12,6 +12,7 @@ use Symfony\Component\Form\Extension\Core\Type\PasswordType;
 
 class PasswordOptionsBuilder extends AbstractOptionsBuilder
 {
+    /** @return class-string */
     public static function getBuilderType(): string
     {
         return PasswordType::class;