+++ src/OptionsBuilder/TimezoneOptionsBuilder.php
@@ -14,6 +14,7 @@ class TimezoneOptionsBuilder extends AbstractOptionsBuilder
 {
     use ChoiceTypeTrait;
 
+    /** @return class-string */
     public static function getBuilderType(): string
     {
         return TimezoneType::class;